requirements_pranjal.txt
absl-py==1.2.0
aiohttp==3.8.3
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
docker-pycreds==0.4.0
einops==0.6.0
eval_type_backport==0.2.2
filelock==3.8.2
frozenlist==1.3.3
fsspec==2022.11.0
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.11.1
idna==3.4
joblib==1.2.0
lightning-utilities==0.4.2
llvmlite==0.41.1
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.3
networkx==3.1
numba==0.58.1
numpy==1.24.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
nltk==3.9.1
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
ply==3.11
protobuf==3.20.1
pydantic==2.10.5
pydantic_core==2.27.2
PyQt5==5.15.10
python-dateutil==2.8.2
python-dotenv==1.0.1
#pytorch-lightning==1.8.5.post0
#pytorch-metric-learning==1.6.3
pytz==2022.7
regex==2022.10.31
requests==2.28.1
sacremoses==0.0.53
scikit-learn==1.2.0
scipy==1.9.3
sentencepiece==0.1.97
sentry-sdk==2.20.0
setproctitle==1.3.4
smmap==5.0.2
sympy==1.13.3
tensorboardX==2.5.1
threadpoolctl==3.1.0
tokenizers==0.10.3
torch==2.4.1
#torchmetrics==0.11.0
#torchvision==0.14.1
tqdm==4.64.1
transformers==4.14.1
triton==3.0.0
typing_extensions==4.12.2
urllib3==1.26.13
wandb==0.19.3
webencodings==0.5.1
yarl==1.8.2
